backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.84.0
outcome==1.3.0.post0
packaging==25.0
pydantic==2.11.5
pydantic_core==2.33.2
PySocks==1.7.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
